+++ lib/tools/coords/_common/dec/logic/dec.dart
