+++ lib/tools/coords/_common/dec/widget/gcw_coords_dec.dart
