+++ lib/tools/coords/_common/lambert/logic/lambert.dart
