+++ lib/tools/coords/_common/mgrs_utm/logic/mgrs.dart
@@ -1,6 +1,6 @@
 import 'package:gc_wizard/tools/coords/_common/logic/coordinates.dart';
 import 'package:gc_wizard/tools/coords/_common/logic/ellipsoid.dart';
-import 'package:gc_wizard/tools/coords/format_converter/logic/utm.dart';
+import 'package:gc_wizard/tools/coords/_common/utm/logic/utm.dart';
 import 'package:latlong2/latlong.dart';
 
 String digraphLettersEast = "ABCDEFGHJKLMNPQRSTUVWXYZ"; //without I and O
