+++ lib/tools/coords/_common/reversewherigo_waldmeister/logic/reverse_wherigo_waldmeister.dart
@@ -1,5 +1,5 @@
 import 'package:gc_wizard/tools/coords/_common/logic/coordinates.dart';
-import 'package:gc_wizard/tools/coords/format_converter/logic/dec.dart';
+import 'package:gc_wizard/tools/coords/_common/dec/logic/dec.dart';
 import 'package:latlong2/latlong.dart';
 
 LatLng reverseWIGWaldmeisterToLatLon(ReverseWherigoWaldmeister waldmeister) {
