+++ lib/tools/coords/geohashing/widget/geohashing.dart
@@ -5,8 +5,8 @@ import 'package:gc_wizard/application/theme/theme.dart';
 import 'package:gc_wizard/common_widgets/buttons/gcw_iconbutton.dart';
 import 'package:gc_wizard/common_widgets/buttons/gcw_paste_button.dart';
 import 'package:gc_wizard/common_widgets/buttons/gcw_submit_button.dart';
-import 'package:gc_wizard/tools/coords/_common//degrees_latlon/degrees_lat_textinputformatter.dart';
-import 'package:gc_wizard/tools/coords/_common//degrees_latlon/degrees_lon_textinputformatter.dart';
+import 'package:gc_wizard/tools/coords/_common/dec/widget/degrees_lat_textinputformatter.dart';
+import 'package:gc_wizard/tools/coords/_common/dec/widget/degrees_lon_textinputformatter.dart';
 import 'package:gc_wizard/common_widgets/coordinates/gcw_coords_output/gcw_coords_output.dart';
 import 'package:gc_wizard/common_widgets/coordinates/gcw_coords_output/gcw_coords_outputformat.dart';
 import 'package:gc_wizard/common_widgets/dialogs/gcw_dialog.dart';
